WIP: snapshot migration
